+++ src/main/java/io/fabric8/maven/docker/config/BuildImageConfiguration.java
@@ -3,6 +3,8 @@ package io.fabric8.maven.docker.config;
 import java.io.File;
 import java.io.Serializable;
 import java.util.*;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 import io.fabric8.maven.docker.util.*;
 import org.apache.maven.plugins.annotations.Parameter;
@@ -79,6 +81,9 @@ public class BuildImageConfiguration implements Serializable {
     private Map<String, String> fromExt;
 
     @Parameter
+    private List<String> cacheFrom;
+
+    @Parameter
     private String registry;
 
     @Parameter
@@ -108,8 +113,12 @@ public class BuildImageConfiguration implements Serializable {
     @Parameter
     private String cleanup;
 
+    @Deprecated
     @Parameter
     private Boolean nocache;
+
+    @Parameter
+    private Boolean noCache;
 
     @Parameter
     private Boolean optimise;
@@ -174,7 +183,16 @@ public class BuildImageConfiguration implements Serializable {
     }
 
     public File getContextDir() {
-        return contextDir != null ? new File(contextDir) : getDockerFile().getParentFile();
+        if (!isDockerFileMode()) {
+            return null;
+        }
+        if (contextDir != null) {
+            return new File(contextDir);
+        }
+        if (getDockerFile().getParentFile() == null) {
+            return new File("");
+        }
+        return getDockerFile().getParentFile();
     }
 
     public String getContextDirRaw() {
@@ -218,6 +236,10 @@ public class BuildImageConfiguration implements Serializable {
 
     public Map<String, String> getFromExt() {
         return fromExt;
+    }
+
+    public List<String> getCacheFrom() {
+        return cacheFrom;
     }
 
     public String getRegistry() {
@@ -280,8 +302,14 @@ public class BuildImageConfiguration implements Serializable {
         return CleanupMode.parse(cleanup != null ? cleanup : DEFAULT_CLEANUP);
     }
 
-    public boolean nocache() {
-        return nocache != null ? nocache : false;
+    public boolean noCache() {
+        if (noCache != null) {
+            return noCache;
+        }
+        if (nocache != null) {
+            return nocache;
+        }
+        return false;
     }
 
     public boolean optimise() {
@@ -293,7 +321,7 @@ public class BuildImageConfiguration implements Serializable {
     }
 
     public Boolean getNoCache() {
-        return nocache;
+        return noCache != null ? noCache : nocache;
     }
 
     public Boolean getOptimise() {
@@ -404,6 +432,22 @@ public class BuildImageConfiguration implements Serializable {
             return this;
         }
 
+        public Builder cacheFrom(String cacheFrom, String ...more) {
+            if (more == null || more.length == 0) {
+                return cacheFrom(Collections.singletonList(cacheFrom));
+            }
+
+            List<String> list = new ArrayList<>();
+            list.add(cacheFrom);
+            list.addAll(Arrays.asList(more));
+            return cacheFrom(list);
+        }
+
+        public Builder cacheFrom(Collection<String> cacheFrom) {
+            config.cacheFrom = cacheFrom != null ? new ArrayList<>(cacheFrom) : null;
+            return this;
+        }
+
         public Builder registry(String registry) {
             config.registry = registry;
             return this;
@@ -497,8 +541,8 @@ public class BuildImageConfiguration implements Serializable {
             return this;
         }
 
-        public Builder nocache(Boolean nocache) {
-            config.nocache = nocache;
+        public Builder noCache(Boolean noCache) {
+            config.noCache = noCache;
             return this;
         }
 
@@ -566,7 +610,10 @@ public class BuildImageConfiguration implements Serializable {
 
         initDockerFileFile(log);
 
-        if (healthCheck != null) {
+        if (cacheFrom != null && !cacheFrom.isEmpty()) {
+            // cachefrom query param was introduced in v1.25
+            return "1.25";
+        } else if (healthCheck != null) {
             // HEALTHCHECK support added later
             return "1.24";
         } else if (args != null) {