+++ src/main/java/io/fabric8/maven/docker/config/ImageConfiguration.java
@@ -25,6 +25,12 @@ public class ImageConfiguration implements StartOrderResolver.Resolvable, Serial
     private String alias;
 
     @Parameter
+    private String stopNamePattern;
+
+    @Parameter
+    private String removeNamePattern;
+
+    @Parameter
     private RunImageConfiguration run;
 
     @Parameter
@@ -69,6 +75,14 @@ public class ImageConfiguration implements StartOrderResolver.Resolvable, Serial
     @Override
 	public String getAlias() {
         return alias;
+    }
+
+    public String getStopNamePattern() {
+        return stopNamePattern;
+    }
+
+    public String getRemoveNamePattern() {
+        return removeNamePattern;
     }
 
     public RunImageConfiguration getRunConfiguration() {
@@ -138,7 +152,7 @@ public class ImageConfiguration implements StartOrderResolver.Resolvable, Serial
         // If there is no explicit run configuration, its a data image
         // TODO: Probably add an explicit property so that a user can indicated whether it
         // is a data image or not on its own.
-        return getRunConfiguration() == null;
+        return run == null;
     }
 
     public String getDescription() {
@@ -192,6 +206,16 @@ public class ImageConfiguration implements StartOrderResolver.Resolvable, Serial
 
         public Builder alias(String alias) {
             config.alias = alias;
+            return this;
+        }
+
+        public Builder removeNamePattern(String removeNamePattern) {
+            config.removeNamePattern = removeNamePattern;
+            return this;
+        }
+
+        public Builder stopNamePattern(String stopNamePattern) {
+            config.stopNamePattern = stopNamePattern;
             return this;
         }