+++ src/test/java/io/fabric8/maven/docker/access/BuildConfigTest.java
@@ -17,6 +17,7 @@ package io.fabric8.maven.docker.access;/*
 
 import org.junit.Test;
 
+import java.util.Arrays;
 import java.util.Collections;
 import java.util.Map;
 
@@ -79,5 +80,20 @@ public class BuildConfigTest {
         assertEquals("0", opts.getOptions().get("nocache"));
         opts.addOption("nocache","1");
         assertEquals("1", opts.getOptions().get("nocache"));
+    }
+
+    @Test
+    public void cacheFrom() {
+        BuildOptions opts = new BuildOptions().cacheFrom(Arrays.asList("foo/bar:latest"));
+        assertEquals("[\"foo/bar:latest\"]", opts.getOptions().get("cachefrom"));
+
+        opts.cacheFrom(Arrays.asList("foo/bar:latest", "foo/baz:1.0"));
+        assertEquals("[\"foo/bar:latest\",\"foo/baz:1.0\"]", opts.getOptions().get("cachefrom"));
+
+        opts.cacheFrom(Arrays.asList());
+        assertEquals(null, opts.getOptions().get("cachefrom"));
+
+        opts.cacheFrom(null);
+        assertEquals(null, opts.getOptions().get("cachefrom"));
     }
 }